Move tree_cache types into fcmp_pp_types.h
